orchestrator/requirements.txt
apache-airflow==2.8.2
apache-airflow-providers-amazon==8.17.0
apache-airflow-providers-google==10.14.0
apache-airflow-providers-apache-beam==5.6.0
